requirements.txt
asgiref==3.4.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.1
django-environ==0.8.1
django-rq==2.5.1
djangorestframework==3.13.1
drf-spectacular==0.21.1
drf-spectacular-sidecar==2021.12.13
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.3.3
MarkupSafe==2.0.1
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.6
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
redis==4.1.0
requests==2.27.0
rq==1.10.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
supervisor==4.2.4
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
